Updated tests for newly hard coded oracle integer type

+++ liquibase-core/src/test/java/liquibase/verify/saved_state/minimumRequiredIsValidSql/createTable/oracle.sql
@@ -4,4 +4,4 @@
 --     type="int"
 -- ], ]
 -- Change Parameter: tableName=person
-CREATE TABLE person (id INTEGER);
+CREATE TABLE person (id NUMBER(10));
