+++ liquibase-core/src/test/java/liquibase/verify/saved_state/minimumRequiredIsValidSql/modifyDataType/oracle.sql
@@ -2,4 +2,4 @@
 -- Change Parameter: columnName=id
 -- Change Parameter: newDataType=int
 -- Change Parameter: tableName=person
-ALTER TABLE person MODIFY id INTEGER;
+ALTER TABLE person MODIFY id NUMBER(10);
